requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
# Pin numpy<2 để tương thích với chromadb 0.4.x
numpy<2
chromadb==0.4.24
pydantic>=2.0
python-dotenv>=1.0
requests>=2.31
pypdf>=3.9.0
python-docx>=1.0.0
rank-bm25>=0.2.2
transformers>=4.41.0
PyQt6>=6.7.0
PyQt6-WebEngine>=6.7.0
langid>=1.1.6
python-multipart>=0.0.9
filelock>=3.12
slowapi>=0.1.9
